Use int(id) instead of long(id). Fixes test_edit_share

+++ omeroweb/webclient/webclient_gateway.py
@@ -1753,7 +1753,7 @@ class OmeroWebGateway(omero.gateway.BlitzGateway):
         """
 
         sh_serv = self.getShareService()
-        sh = sh_serv.getShare(long(oid))
+        sh = sh_serv.getShare(int(oid))
         if sh is not None:
             return ShareWrapper(self, sh)
         else:
